PM / Domains: Allow domain power states to be read from DT
This patch allows cluster-level idle-states to being soaked in as
generic domain power states, in order for the domain governor to chose
the most efficient power state compatible with the device constraints.
Similarly, devices can register power-states into the cluster domain, in
a manner consistent with idle-states.

This is a attempt to address device-retention states for devices that
are not hooked to runtime-pm, but feature a retention state handled by
the same firmware that handles idle-states. For instance a L2 caches.

With Juno, in this example the idle-state 'cluster-sleep-0 ' is known
from each cluster generic domain, as the deepest sate.

cat /sys/kernel/debug/pm_genpd/*

Domain             State name        Enter (ns) / Exit (ns)
-------------------------------------------------------------
a53_pd               cluster-sleep-0      1500000 / 800000
a57_pd               cluster-sleep-0      1500000 / 800000

domain                      status pstate     slaves
/device                                      runtime status
-----------------------------------------------------------------------
a53_pd                          on
/devices/system/cpu/cpu0                            active
/devices/system/cpu/cpu3                            suspended
/devices/system/cpu/cpu4                            suspended
/devices/system/cpu/cpu5                            suspended
/devices/platform/D1                                suspended
a57_pd                          cluster-sleep-0
/devices/system/cpu/cpu1                            suspended
/devices/system/cpu/cpu2                            suspended

Signed-off-by: Marc Titinger <mtitinger+renesas@baylibre.com>
Signed-off-by: Lina Iyer <lina.iyer@linaro.org>

[Lina: removed dependency on dynamic states, simplified initalization,
added of_pm_genpd_init() API]

[Marc:
- allow for state_count == 0 upon calling state allocation routine
the first time (when the DT contains no domain-idle-states) to not
fail pm_genpd_init
- remove the check for a compatible for power-states, as we have
  a phandle to an idle-state already
- sort out the idle-state to genpd-state property conversion.
- squash Allow domain states to hold additional state related data in a u32
value. This may be used by the platform driver.
- squash : Domains that have expensive suspend resume operations, may require a
certain amount of time be spent in an idle state to reap the power
saving benefit. Such domains, may provide the residency requirement for
a domain state. Read the residency for a domain state from the DT. A domain governor may
use this information in determining the state for the domain.
- make the state init dynamic again, use a simply chained list of states.
]